Added interactive functionality
The script now works with the -i flag to start interactive mode in which
the script asks for each file if you want to link it or not.

Also changed how the root option will work, a regular user can run the
command and the script should check if it needs to use sudo or not (if
root ran the script).