Vim auto-formatting updates
Turned off auto-formatting for text by default because this causes newlines to
be inserted at &textwidth when editing existing wrapped lines, which turns an
existing soft-wrap into a hard-wrap.

When I turn wrapping on I often desire a soft-wrap, so setting
`formatoptions-=t` will cause a soft-wrap. For a hard-wrap this will need to be
reversed.