Added more system config files
10-network.rules: this udev rules file will rename my external Alfa wireless
card interface to "alfa". This has not been tested with multiple cards and will
probably break something if two of the same card are plugged in.

910-rat.conf: this file remaps the buttons on a Saitek Cyborg R.A.T.7 Mouse so
that the horizontal scroll wheel will work.

Both files have comments that say where in the system they should be put to work
properly.