Save the time the plugins were auto-updated
This ignores if the user manually runs :PluginUpdate themselves but I don't
think this is a major issue. This function will automatically run Vundle's
:PluginUpdate if it hasn't done so in a week or more, and it will only do so on
Friday Saturday or Sunday.

The only caveat is that it is currently hooked into the VimEnter autocmd, so
that the checks will only occur when Vim is opened, but this can easily be
changed if more frequent checks are necessary.