Bug: saving jumps the cursor
The autocmd to remove trailing whitespace has a side effect of jumping the
cursor around when saving a file. This is a work around causing the cursor to
jump back, but on subsequent saves the cursor jumps back and forth between it's
last known positions.

A function is probably more suited to solve this.