Added bin/activeNetDev.sh
Now the `local-ip` shell alias can get the active network device no matter what
it is named. It currently depends on `nmcli`, this should probably be changed to
some sort of `ip` command to make it more portable, but for my setup right now
it is fine.

Note: the shell alias will probably fail if more than one network device is
connected.