Vim: set html to use syntastic in passive mode
This is because html files for different projects such as Angular applications
or ejs templates will throw errors because of missing DOCTYPE strings or body
and head tags. This is a temporary solution, I would like to run some sort of
detection here to check if this is regular html or not and conditionally set
passive mode.