Update Git push.default script and filter
The filter will now comment out the push.default option so that when the file is
pulled it will not cause git to error on every command because of invalid
options, also updated the `bin/hasGitPushSimple.sh` script to account for the
comment.

This update should make it play nice with the git hook that will run
`bin/hasGitPushSimple.sh` whenever `gitconfig` updates.

The filter will cause a bit of a problem if the line is already commented out,
but this shouldn't normally be the case.