Vim: Move typo autocorrection out of vimrc
Not everything that I edit in Vim needs typo autocorrection, but it is also too
small a feature to need a separate plugin. So I created a new directory under
`~/.vim/` that I called custom. I then moved typo autocorrection to it's own
vimscript file in this directory and for each filetype that could need typo
autocorrection I have sourced this script.

Also moved all gitcommit filetype settings to the `~/.vim/after/` directory to
avoid confusion or conflicts with the system gitcommit filetype settings.