Vim: Remove CtrlP and update grep commands
Remove CtrlP (and CtrlPFunky) because I did not use them enough to need them.

Add DidYouMean, which will cause vim to ask what to do when opening a new file
that is a partial match of one or more files in the current directory.

Also setup the quickfix window to open after grep commands, and setup a leader
mapping for switching buffers.