Vim: Make tex compile handle beamer files
When running the s:MakePDF() function for LaTeX files it will now check
for the string `\documentclass{beamer}`. If found it will run `pdflatex`
again to account for any navigation or tables of contents, then it will
also remove all the other files types that generating a beamer
presentation produces that a regular document doesn't.