Vim: Change markdown and tex 'compile' mappings
The mappings now make use of 'makeprg'. This makes the mappings much
cleaner.

As a side effect this means that outputs are now put in Vim's current
working directory. This isn't that big a deal, and behaves how the
mappings for C/C++ files work with the ':make' command.

Also added detection of `rubber` to the TeX mapping.