Vim: Fix enter mapping removed after using cmdwin
The cmdwin enter mapping fix was created before the mapping to make
enter a shortcut to command mode, when that mapping was created the
cmdwin options were skipped over. This now behaves as expected but a
more reliable solution is desirable, maybe something that saves what
enter was mapped to before and then replaces it.