bspwm: Add multi-monitor and click to focus
The multi-monitor support should assign three desktops to an external
monitor mirroring the last three desktops of the internal monitor. This
is untested and I do not know if it will automatically detect an
external monitor.

Also this adds click to focus, now the mouse can be used to focus
windows/monitors.