Git: Update post-merge hook to fix status issue
Previously when run the post-merge hook would work as intended but
would leave git thinking the repository was "dirty"; however because of
the filter in `.gitattributes` the `git diff` would be empty.

This change forces git to refresh the index regarding the gitconfig file
after running `bin/hasGitPushSimple.sh` so that git can realize there
are no changes to the file.

Also updated `README.md` regarding my git configuration and the
post-merge hook specifically.