install.sh: Skip files already installed this run
This allows for groups of files to depend on the same file (such as the
Bash and Zsh groups depending on shared shell environment configs.

This change keeps a list of all files linked in a temporary file, which
allows it to avoid attempting to install files a second time.