install: Rewrite how dependent groups work
With the upcoming merge of the bspwm branch (my bspwm configuration) it
will be necessary to have a system in place where a dependent group can
include files in both the home and config directories. This setup was
not possible with the previous implementation.

This moves where the definition of targets is done to where the lists of
files are specified, so each file keeps track of where it is supposed to
be installed and how.

This increases repeated code in some ways as each file will need to
specify a lot of repetitive information and each group will need to
repeat the install loop.

The install loop repetition is due to the fact that bash does not have
multidimensional arrays; although I simulated one by having each file be
a space delimited list so I may yet be able to reduce the code by doing
something similar with the groups.

Also: added boolean variables to keep code readable and uniform, and
changed the 'installed this session' check to be more of a
attempted/checked this session and fixed a small lookup bug where
substrings were matched creating a false positive.