Vim: Add fix for Background Color Erase issue
An issue with Vim's Background Color Erase causes the terminal color to
bleed through Vim when clearing sometimes. I most noticed the issue when
running the `:highlight` command with no arguments. According to [1] the
issue is usually noticed in Tmux or Screen environments; however, I
noticed it in termite and a recent reddit post [2] displayed the same issue in
gnome-terminal so it may be a lib-vte issue as well.

[1]: (https://sunaku.github.io/vim-256color-bce.html)
[2]: (https://www.reddit.com/r/vim/comments/5uulam/background_color_not_all_the_same/?ref=share&ref_source=link)