Vim: Fix command abbreviations for 'B' and 'E'
Previously these abbreviations would be applied in some inappropriate
situations such as `:help B` capital be was desired but it was
translated to `:help b` before executing.

This fix makes sure the abbreviation is only applied when 'B' or 'E' are
used on their own as the start of a ':' command.