Vim: Add Xclip command to make copying out easier
This makes it easy to copy text out of Vim when it is compiled without
the 'clipboard' options.

The command takes in a range (defaulting to the current line) and
filters it through the given command. In this case `tee` is used so that
the text will not be removed from Vim and can still be piped to xclip.