Vim: Fix cmdwin enter mapping clobber
The way the mapping was handled upon exiting the cmdwin the enter key
would be mapped to a colon (':') even if it wasn't previously mapped to
that. Some filetypes such as manpages and help files have enter mapped
to a tag search.

This makes use of a function so whatever the enter key is mapped to can
be saved in a buffer local variable and restored upon exiting the
cmdwin.