Vim: Add LaTeX math mode highlight to Markdown
Markdown files can use underscores to define italics, but if LaTeX math
mode is used underscores will no longer start italics.

This adds a simple LaTeX math mode syntax region to Markdown files. This
region is started by at least one space and a dollar sign ('$') and
ended by a dollar sign and optional spaces (because the region may end a
line or a sentence.

Highlighting was defined to normal within the region and linked to the
Type group for the region delimiters (dollar signs).

Also small spellfile update.