sxhkd: Fix volume keys to only work when not muted
The volume keys had been able to change volume when the system was
muted, leading to sometimes unmuting into a very loud system.

This makes the volume keys only change the volume when the system is
not currently muted. This may pose an issue if multiple audio outputs
are connected.