Git: Clean up alias printing
This makes printing aliases look much better; multiple spaces due to
splitting the alias across line get squashed to one space and the 'fmt'
utility is used to wrap and indent long lines for readability.

It was also necessary to escape a newline in the 'contributions' alias
in order to avoid printing a newline, this seems to have not affected
the functionality of that alias.