Git: Simplify git push.default setup
The difference between upstream and simple isn't really enough to
warrant all this configuration. Upstream has been set as the
push.default value for its backwards compatibility with older versions
of git.

It was a neat experiment getting all this to work, but it's not
necessary and much cleaner to get rid of it all.