Git: Add mkmailmap alias and tweak authors alias
This adds an alias that creates a `.mailmap` file in the current
repository by going through the shortlog and adding every author's name
and email to the `.mailmap` file. From there the file should be edited
to create an appropriate map of author's emails to their name.

This also makes the authors alias print emails.