Git: Make 'whoami' alias ignore 'useConfigOnly'
The 'useConfigOnly' setting (although included in the 'user' section) is
not really part of who a user _is_ and so doesn't need to be shown in
the 'whoami' alias. This adds a sed expression to remove that line.

This commit also combines sed expressions in the 'alias' alias saving an
unnecessary pipe.