Vim: Favor latexmk for building latex files
The latexmk wrapper has better support for bibliographies and still is
as simple as rubber to call. This updates the section that sets up the
F5 keybinding for building to favor latexmk over rubber and rubber over
calling pdflatex directly.

This also adds other output files to cleanup which are created from
bibliography entries.