Vim: Add statusline switching autocmds
The statusline looses highlighting for inactive windows, I believe this
is because User highlight groups (`:help hl-User1`) were used to
highlight it.

This adds functions to set a 'fancy' statusline and a 'simple' one. The
'simple' one doesn't use highlight groups and will rely on the
'Statusline' and 'StatuslineNC' highlight groups. It also adds autocmds
to switch between the 'fancy' and 'simple' statuslines when entering or
leaving a window.