Replace terminal-colors.py with term-colors alias
The 'terminal-colors.py' script is a large chunk of code that I do not
own, while it is perfectly fine for me to distribute it I don't feel
that it belongs here and I can get the same functionality with an alias
found on the ArchWiki.

This adds the 'term-colors' alias to 'shell_aliases' in order to replace
the functionality of 'terminal-colors.py'. This also applies some
shellcheck fixes to 'shell_aliases'.