Vim: Replace qmake syntax and ftdetect with plugin
This code is also code that I don't own so I feel it does not need to be
here. This replaces it with a single line in my 'vimrc' to install the
files as a plugin. Plus as a side effect, I don't have to make changes
if the maintainer ever updates it.