Vim: Add new superupdate config option
When performing a fresh install of the Vim configuration Vundle would
end up running twice, once to install everything and once as
Superupdate's first check.

This takes advantage of a new Superupdate feature to skip checking on
the first run (because we know the plugins are up to date at that time).