git: Add git-summary command
This will eventually take the place of the contributions alias. Turning
it into an extension program will make it cleaner and easier to read,
and hopefully more robust.

This gets the initial setup of the script done as well as adds zsh
completion for it. It still needs to check authors against a mailmap
file if one exists and print their contribution summary.