vim: Replace vundle with vim-plug
This replaces Vundle with vim-plug for a simpler config, greater control
over plugin loading and better performance. Vim-plug itself is ignored
in the gitignore.

It also adds chrisbra/Colorizer and only loads it for css and html
files, Yggdroot/indentLine is also only loaded for html, xhtml and xml
at this time.