shell: Make vim MANPAGER cleaner
Using vim as the MANPAGER works, but when jumping between man pages
using vim's tag stack it's impossible to get back to the first man page
because it was opened from stdin.

This does not solve that issue, but it does move the stdin processing
from the MANPAGER variable to vim itself, which cleans up the look of
everything. This seems to work identically to before.

    - cleanup idea from [1]
    - the `let $MANPAGER=''` line is taken care of by sourcing
      $VIMRUNTIME/ftplugin/man.vim

This also cleans up how executables are checked for in shell_env which
actually lets lesspipe get sourced correctly.

[1]: https://murukesh.me/2015/08/28/vim-for-man.html