vim: Add autoload script for saving stdin manpage
When using vim as the system manpager the first manpage is read from
stdin and not added to the tag stack, which becomes annoying when
jumping between other man pages as you cannot get back to the first one.

This adds an autoloaded function to setup the stdin and save the manpage
to a temp file. Then it adds an autocmd to clean up the temp file when
closing vim.