vim: Update many bits
This includes a bunch of small updates.

It finishes the highlights for QuickScopePrimary and
QuickScopeSecondary.

It adds scss (SASS) files to Colorizer.

It cleans up and comments the numbering_group autocmds. Just combines
two autocmds that trigger on the same event and comments better what the
section is for.

It adds a few bits to the ToggleHexMode function. This makes the
function better, but it could probably still use some updating. The
`setlocal nomodified` bit probably won't work very well if actually
modifying the hex code.

It cleans up the ToggleDiffMode function. There was no reason to use
execute when `diffoff` and `diffthis` are perfectly good commands.

It makes the `set list` leader mapping only effect the current buffer.