npm: Add npmrc and global npm config
This adds an npmrc to install global packages to '~/.npm_pkgs'. It adds
'~/.npm_pkgs/bin' to PATH and '~/.npm_pkgs/share/man' to MANPATH. It
adds npmrc to 'install.sh' and fixes a typo in the 'promptYN' function
of 'install.sh' at the same time.

This also removes the npm pacman hook because it would no longer work
with how packages are installed.