vim: Add LaTeX insert abbreviation for infinity
I often forget that in LaTeX the infinity symbol is `\infty` and not
`\infinity`. This adds an insert mode abbreviation to fix this; however,
Vim does not allow '\' to be included in abbreviations, so using the
solution from [1] I created a function that will abbreviate if the word
was preceded by a '\' character.

[1]: https://stackoverflow.com/a/1680834