bspwm: Add dunst config
This adds a mostly default dunstrc, copied from
'/usr/share/dunst/dunstrc'. It already provides better notifications
than without a dunstrc. It adds this to the install script and sets up
Vim do detect it as dosini syntax, which is loosely correct, dunst
itself seems to recommend the cfg filetype.