ranger: Remap gm to use /run/media/USER
The default for ranger's 'gm' is to go to '/media/' but the udisks2
convention is '/run/media/USER'. This remaps 'gm' to use the udisks2
convention.

This makes use of ranger's `fm.cd` which expects a python path object
and python's `os.path.expandvars` which will expand environment
variables. Doing it this way allows this 'rc.conf' to be used regardless
of username differences on any hosts this may be installed on.