vim: Change <leader>m mapping to make
This adds a new mapping for '<leader>m' to call ':make' and renames the
old mapping to '<leader>u/U'. The previous mapping was on 'm' as an
mnemonic for 'mark'; however, 'make' is a much more useful command and
more fitting of the 'm'. The 'u/U' is used as a replacement because of
the 'underline' mnemonic.