zsh: Make tmux prompt bold when attached
The tmux portion of the prompt shows the number of tmux sessions on the
machine. The number isn't exactly helpful, this makes the prompt portion
bold when attached to a tmux session (based on checking '$TMUX'). Now it
is easy to see not only how many active tmux sessions there are but
whether or not one is currently attached.