vim: Make Xclip command not modify the buffer
The previous implementation of the 'Xclip' command made use of a filter
which modified the buffer text, it piped it out to `tee` which returned
it unchanged but Vim still registered it has a change.

This makes use of the 'write' command to write directly to the stdin of
an external program. This update also removes the need for `tee`.

Also add the 'node_modules' folder to 'wildignore'.