vim: Simplify keywordprg settings and fix noremaps
Previously I though 'keywordprg' could only be external commands, but it
can also be internal commands such as ':Man' or ':help'. This changes
any mapping of 'K' to a 'setlocal keywordprg=<something>'.

Also, any instance of 'noremap' has been changed to 'nnoremap' because
that is most likely what I intended.