vim: Move functions to autoload
Most of the functions in my vimrc are for things I don't use all the
time, which makes them the perfect candidates for autoload. This moves
all but the statusline functions to files in the autoload directory.
Additionally this implements wincent's deferred starting as demonstrated
in his screencast about profiling and optimizing [1].

[1]: https://www.youtube.com/watch?v=wQ9uB8I0cCg