bspwm: Add polybar urgent highlights
This adds an orange highlight for bspwm workspace names with urgent
windows to the polybar config. It also changes the keybindings for
resizing floating windows to something that feels more natural to me. I
haven't tested the urgent window highlighting but it hasn't messed
anything up either so I don't see a reason not to include it right now.