vim: Set LaTeX makeprg only for the main file
Running a LaTeX compilation from an included file rather than the main
file produces errors. This is a quick solution to "smartly" set the
makeprg and F5 binding for only the main file. There are foreseeable
situations in which this will break (namely if a main file is started
from scratch rather than one of my templates), but for my main use case
right now it seems to work fine.

This also adds spaces around the '==?' operator in the BSlashReplace
function for readability.