bspwm: Add btops to manage workspaces
This adds btops [1] to dynamically manage bspwm workspaces. This means
that all the bits in bspwmrc dealing with workspaces are no longer
needed.

This also updates the polybar config quite a bit. The battery and volume
percentages are prefixed with 'b:' and 'v:' respectively, there is now
an alternate date format activated by clicking and workspaces are
displayed differently. To go along with btops empty workspaces are
displayed as a '+' (since there should only ever be one empty one. I've
also changed the underline to represent the focused workspace and color
to represent occupied workspaces. The color of the monocle flag was also
changed to red.

[1]: https://github.com/cmschuetz/btops