xresources: Move colors into separate file
This moves the colors into a separate file called
'tango-colors.xdefaults' because those colors mostly follow the tango
colorscheme. This also adds 'bsmith-colors.xdefaults' which is a working
title for a working colorscheme. I'm not sure that I'm happy with it,
and I think I need to use it for a while to figure that out.