bspwm: Change rofi launcher to only show GUI apps
This changes the launcher keybinding in sxhkd to use rofi's 'drun' mode
which will only show executables that have desktop files (i.e. they are
GUI programs). Most of the time that I open rofi I intend to launch a
GUI program so this is a better mode to use than the default 'run' mode.