git: Add 'git nuke' alias
This adds an alias which will remove the '.git' folder of a repository,
that is essentially nuking the git information. I also made sure that
`git rev-parse` returns without error before running the `rm` to
hopefully ensure that no unintended `rm`s happen.

The alias isn't super helpful, but one that I've been considering adding
for a while now and had a use-case for today, so it's here now.