shell: Add vim helper to format man-like input
In the $MANPAGER variable vim will call a new function that I just added to do
some basic formatting (`col -b`) if the filetype isn't already set to 'man'.
This makes things that aren't 'man' that interact with the system pager like
'man' does work. For example, some python utilities have a help subcommand that
prints a manpage using the system pager, however they do not set $MAN_PN like
'man' does so they miss my initial manpage detection.

This will set the wrong filetype if something uses the pager for input that
isn't man-like in syntax, right now I consider this to be a minor drawback as I
don't know of any examples of this in my current workflow.