vim: Fix manpage#Prettify() finish error
Fixes error introduced in: c217dcaa36db8bda1916a4e194c1b6c75643af85

When viewing a regular manpage the manpage#Prettify() function should do
nothing. It produced an error that 'finish' was used outside of a
sourced file. To fix this I changed the conditional to the opposite and
wrapped the function content.