shell: Add ssh-ident aliases
This sets up for me to use ssh-ident to manage my ssh-agent (that is,
ssh-ident will automatically invoke ssh-add for keys when they are
needed). This commit adds aliases for ssh, scp, sftp and rsync so that
they each use ssh-ident instead of plain ssh.

It also adds a line to my zshrc so that ssh-ident will use the same
completion as ssh.

At some point I may want to add a ~/.ssh-ident file to configure the
ssh-add timeout.