shell: Move ssh-agent output
This moves the output from ssh-agent from ~/.ssh-agent-output to
~/.ssh/agents/agent-$USER-$(hostname) because this is where the
ssh-ident tool expects it. With it in the previous location ssh-ident
was spawning a second ssh-agent, this prevents that so there is only one
ssh-agent running.

This also adds an alias for git to make sure it also uses ssh-ident
instead of ssh.