shell: Add line to make ssh agent directory
When I setup ssh-ident it had already created the $HOME/.ssh/agents/
directory for me, this adds a line to shell_env that will create that
directory automatically.

Note: if $HOME/.ssh/ doesn't exist when this is run it will be created
with default umask instead of 0700 because when mkdir is run with '-m'
and '-p' the mode only applies to the deepest directory created
(according to shellcheck 2174).