shell: Remove ssh-ident in favor of config option
This removes the ssh-ident aliases in favor of using the AddKeysToAgent
option in the ssh config file. So this also adds an ssh/config to the
repo and to the install script.

The install script makes sure ~/.ssh exists and links the config file
into it rather than linking an entire directory like the rest of this
repo. Additional reasoning for this choice has been included in the
README.

This also adds an ftdetect script for vim to recognize ssh config files
in this repo as well as ~/.ssh/config.local files that I may create on
machines I use.