git: Reorganize aliases
This formats the comments better in my gitconfig, as well as moving
aliases to be grouped by similar function. That is, all the 'English
words to list things' aliases are in a group, 'question aliases' are
together, and the two log by day aliases are together and now next to
the 'new' aliases. Leaving miscellaneous aliases for the end. This new
ordering seems to flow better than the previous semi-alphabetical. All
comment characters are now followed by a space for readability.

The only behavioral change is to the 'whatis' alias, which used to have
the '-s' option effectively making it the same as git log for a single
commit, this just makes it an alias to git show, which includes the
diff.

This also adds a newline to my git commit template file that I have been
wanting for a while. The space between the 'Details' line and the
comments will make it easier to format paragraphs without getting the
comments caught up in the formatting.