Change mirrorlist hook to symlink
I found an error in the mirrorlist hook where I forgot to add in output
redirection to the call to `command` here but didn't forget in the
script in bin. This solves that error by making this a symlink to the
same script in bin. There is no reason to save two versions of the same
script in this repo.