ranger: Add ~/Downloads mapping
This adds a mapping for ranger to use gj to go to the ~/Downloads/
folder. I chose j to match web browser patterns of Ctrl+j to open
downloads.

Side note: this also changes the date format of my updated: line to use
YYYY-MM-DD, I plan to change to this format as I update other files.