vim: Add additional syntax to git blame
This changes the git blame buffer filetype from just 'git' to a new
custom 'gitblame' and adds a syntax file for it. The syntax just
includes the git syntax and adds some highlighting for a few things I
thought were missing.

New highlights: author and date is highlighted instead of left white,
hashes beginning with a '^' are now highlighted like the rest of the
hashes and empty not yet committed hashes are highlighted with NonText
to draw attention to them.