git: Make log (and aliases) respect mailmap
This adds the log.mailmap option to make git log commands respect the
.mailmap file like the shortlog commands. This also updates aliases
relying on git log to print authors using '%aN' and '%aE' over '%an' and
'%ae' in order to also respect .mailmap files.

In addition to that an issue where the lines and line-authored aliases
don't work with files that have spaces in them is fixed. This uses the
git grep '-z' option to separate items with nulls and the xargs '-0'
option to split on those nulls. Now this will only be broken if a
filename has a null in it (which should hopefully be less common than
spaces).