vim: Make git helpers automatically delete buffers
Changing the bufhidden settings to delete automatically wipes out the
buffers when closing them so that they can be reopened (or another git
blame or log can be opened) without errors.

This also removes spaces from the buffer names as older versions of Vim
complained about that, as well as adding git back to the blame filetype
to take advantage of settings and mappings already builtin to Vim.