vim: Fix python auto-wrap, update PYTHONPATH
This makes sure python files will not auto-wrap at textwidth, wrapping
python needs to be done more carefully than this.

This also includes a quick hack to make sure the current directory (as
far as Vim is concerned) is in the PYTHONPATH, which makes syntax
checking custom modules much happier. This is probably a symptom of a
poorly configured python environment, but it works reasonably well for
now.