git: Make graph deal with long commit messages
This makes my graph alias wrap lines so that really long commit message
(first lines) don't make the output messed up. It also better supports
smaller screens/terminals now.

This also adds the alias 'shortgraph' to show an even shorter output.
The idea was to show less information to minimize the need for wrapping
and get a better idea of the flow of the project.