shell: Remove pgrep dependency for cygwin
When using this environment in Cygwin pgrep doesn't exist and causes ssh
to set itself up incorrectly. This replaces pgrep with a ps grep
combination that produces the same effect.

Shellcheck actually suggested that I replace this change with pgrep, but
I disabled the check because I was specifically avoiding pgrep here.