git: Add automatic ctags hooks
This adds git template hooks for repositories to automatically generate
and re-generate ctags. This is following suggestions from the Vim advent
calendar [1] and Tim Pope's blog post [2] (where much of it was actually
taken from).

Some slight adjustment was made, since I no longer use fugitive, I moved
the generated tags file to the root of the repository (which may need to
be tweaked) instead of leaving it in the git directory. This is so the
whole setup doesn't require more work in my vim configuration.

[1]: https://vimways.org/2018/you-should-be-using-tags-in-vim/
[2]: https://tbaggery.com/2011/08/08/effortless-ctags-with-git.html