Fix bspwm screen lock
The interaction between i3lock-fancy and scrot changed somehow. There is
an error in ImageMagick convert about the file header not being an
image.

This avoids the error by letting i3lock-fancy take the image itself. A
side effect is that the screen lock is now possibly a small bit slower.