vim: Add WSL compatibility and cleanup
This adds a fix for the cursor changing shape when running vim in WSL
(Windows Subsystem for Linux).

It also goes through and removes many unneeded comments and empty lines
for some general cleanup.

Also, editorconfig no longer requires the external core, there is a
builtin vimscript core, so a couple conditionals were removed for that.