vim: Add rooter for better project search
This adds vim-rooter so that I can lookup the root project directory,
based on the presence of .git/ directory (or other configurable
files/directories) to improve project-wide searching.

The grep abbreviation is updated to use vim-rooter to trigger a grep
from the project root.

This also includes an abbreviation to make :helpclose more like :cclose.