Add raise script
This is a small script to use for global hotkeys to focus a specific
window. This script focuses the window if it exists somewhere else it
calls the given name to launch the program. There are possibly cases in
which this doesn't work (i.e. window name is not program name or
something), but for my needs so far this is working.