vim: Replace neocomplete for v8.2.1066+
Neocomplete is no longer compatible with Vim 8,2,1066+, so if that
version or higher is being used this will replace neocomplete with
supertab.

Shougo's deocomplete is the recommended replacement for neocomplete, but
it's installation is much more involved (requires two other plugins for
support as well as a python neovim compatibility package).

So I opted to go for a much simpler autocomplete plugin. It's far less
automatic, but I most likely won't need much autocompletion help. This
can also get me more capable of using Vim's built in completion
commands.