git: Add multiple improvements
This allows git to take a local config file matching most of my other
configs. The matching vim filetype detection is added along with this
change.

Using a local config file I setup some https credential caching for a
specific repo and found that broke the 'whoami' alias, the fix was to
only match user at the beginning of config items.

This allows for a global mailmap file instead of creating and ignoring
one for multiple repos.

Add the 'graph-branch' alias to show graph view for only the current
branch instead of the whole tree, and the 'merged-branches' to show
which branches have been completely merged to master.

Also add the pre-push hook from
https://github.com/da-x/misc-gitology/master/hooks/pre-push to not allow
commits that start with "WIP" or a few other things to be pushed.