bin: Fix raise on latest Gnome desktop
This script was not properly activating the window it was supposed to,
it's not clear if this was a gnome or wmctrl update on Arch that caused
this.

Either way it seems we can just change the 4 to a 3 (which may target a
different or the same field in the wmctrl output).