vim: Fix highlighting of git diffs in commit message
The highlighting of the diff included in the editor from `git commit -v`
requires trailing spaces. Previously I had disabled my custom trailing
spaces removal and thought this fixed, but the issue came back in
repositories using editorconfig (set to remove trailing spaces).

This adds a disable for editorconfig in the buffer as well.