bin: Tweak reflect_mirrors.sh and add prevent sleep
This tweaks the reflect_mirrors.sh script to only pull the 10 fastest
mirrors (this seems to behave better than 200, and really who needs that
many mirrors).

This also adds a script that can be used to configure pam to prevent a
machine from sleeping if it has active ssh connections (so that sleep
doesn't kill ssh sessions).