vim: Update version/patch checking
The latest vim seems to behave unexpectedly with the
`has('patch-X.X-XXX')` format, this was noticed with the
`listchars:space` setting not displaying even though the current vim
(9.1) was way later than the specific 7.4 patch needed.

This avoids the issue by changing to the older method of checking
version number and then `has('patchXXX')` format. It also makes use of a
couple scriptlocal variables to avoid duplicating complex version check
logic.